@@ -13,32 +13,27 @@ license = Apache-2.0
 classifiers =
   Development Status :: 4 - Beta
   Intended Audience :: Developers
-#   Programming Language :: Python :: 3.6 ## numpy >= 1.20.0 dropped support for python 3.6
-  Programming Language :: Python :: 3.7
-  Programming Language :: Python :: 3.8
+  Programming Language :: Python :: 3.9
   Topic :: Scientific/Engineering :: Artificial Intelligence
   Topic :: Scientific/Engineering :: Image Recognition
 
 [options]
-python_requires = >=3.7,<4
+python_requires = >=3.9,<4
 zip_safe = False
 include_package_data = True
 packages = find:
 install_requires =
     pillow >8.0.0,<9
-    torch >=1.9.0,<1.11
-    torchvision >=0.10.0,<0.12
-    fastcore >=1.3.0,<1.4
+    torch == 1.13.1
+    torchvision == 0.14.1
+    fastcore >=1.3.8,<1.5
     tqdm >=4.49.0,<5
     opencv-python >=4.1.1,<5
     albumentations >=1.0.3,<1.1
     resnest >=0.0.6b20201125,<0.0.7
-    effdet >=0.2.1,<0.3
+    effdet >=0.4.1,<0.5
     sahi >=0.11.1,<0.12.0
-    yolov5-icevision >=6.0.0
-    importlib-metadata >=1;python_version<"3.8"
-    ipykernel >=4.10.1,<6
-    dataclasses ==0.6
+    yolov5 ==6.2.0
     loguru >=0.5.3
     rasterio>=1.3a1, <2
 
@@ -48,8 +43,8 @@ inference =
     icevision
 
 all =
-    fastai >=2.5.2,<2.6
-    pytorch-lightning >=1.5.0
+    fastai >=2.7.10,<2.8
+    pytorch-lightning >=1.5.0,<2
     wandb >=0.10.7
 
 dev =
@@ -60,7 +55,7 @@ dev =
     mkdocs >=1.1.2,<2
     mkdocs-material >=7.3.1,<8
     mike >=1.0
-    jupyter >=1.0.0,<2
+    notebook >=6.0.0,<7
     Sphinx >=3.1.0,<4
     pytest-cov >=2.10.1,<3
     flake8 >=3.8.3,<4
