@@ -4,7 +4,7 @@ from icevision.all import *
 
 @pytest.fixture()
 def records():
-    imgs = [np.ones((2, 4, 3), dtype=np.float), np.ones((3, 2, 3), dtype=np.float)]
+    imgs = [np.ones((2, 4, 3), dtype=np.float32), np.ones((3, 2, 3), dtype=np.float32)]
 
     records = []
     for img in imgs:
@@ -20,7 +20,7 @@ def test_img_pad_stack(records, pad_value):
     tfmed_records = tfms.batch.ImgPadStack(pad_value=pad_value)(records)
     imgs = np.asarray([record.img for record in tfmed_records])
 
-    expected = np.ones((2, 3, 4, 3), dtype=np.float)
+    expected = np.ones((2, 3, 4, 3), dtype=np.float32)
     expected *= np.array(pad_value).reshape(-1)
     expected[0, :2, :4, :3] = 1
     expected[1, :3, :2, :3] = 1
